Revert "Feat/85 create check servername"
