--- inc/CommandHandler.hpp
@@ -1,17 +1,0 @@
-#ifndef COMMANDHANDLER_HPP
-#define COMMANDHANDLER_HPP
-
-#include <string>
-#include <vector>
-
-class CommandHandler {
-public:
-  void handleCommad(const std::string &command, User &user);
-
-private:
-  void parseCommand(const std::string &commad);
-  void executeCommand(const std::string &commandName,
-                      const std::vector<string> &params, User &user);
-};
-
-#endif
