--- inc/Replies.hpp
@@ -1,3 +1,0 @@
-class Replies {
-public:
-};
