+++ inc/Server.hpp
@@ -1,62 +1,90 @@
 #ifndef SERVER_HPP
 #define SERVER_HPP
 
-#include "Channel.hpp"
-#include "User.hpp"
-#include <map>
-#include <netinet/in.h>
-#include <poll.h>
-#include <set>
-#include <string>
-#include <vector>
-#include <cstdlib>
-#include <iostream>
+#include "ircserv.hpp"
+#include "util.hpp"
+#define MAX_MESSAGE 512
 
-#define SERVER_NAME "localhost"
-#define PASS_MAX_LEN 32
-#define SERVER_MAX_LEN 63
+// color
+#define NC "\x1B[0m"
+#define RED "\x1B[0;31m"
+#define GRN "\x1B[0;32m"
+#define PPL "\x1B[0;34m"
+#define CYN "\x1B[0;36m"
 
-class Server {
-public:
-  Server(int argc, const char *argv[]);
-  ~Server();
-  void run();
-  void changeNickname(const std::string &before, const std::string &after);
-  const std::string &getServerName() const;
-  const std::string &getNickHistory() const;
-  void setNickHistory(const std::string &nick);
+// Command ===================================================
+// status
+#define ERROR 0
+#define NICK 1
 
+class Command {
 private:
-  std::string
-      serverName; //    > 1.1 サーバー
-                  //    サーバーは名前で一意に識別されます。この名前は、最大63文字（63）の文字です。サーバー名で使用される可能性のあるもの、および使用されないものについては、プロトコル文法ルール（セクション2.3.1）を参照してください。
-
-  std::string password; // 32 自分で決めた
-
-  short port; //
-
-  int sfd;
-  struct sockaddr_in _addr;
-  std::vector<struct pollfd> _pollFd;
-
-  void checkServerName(const std::string &serverName) const;
-  void checkArgc(const int argc) const;
-  void checkArgv(const char *argv[]);
-  void checkPortNum(const short port) const;
-  void checkPassword(const std::string &password) const;
-
-  // std::map <const size_t clFd, std::string nickname> fdToNickname;
-  std::map<const size_t, std::string> fdToNickname;
-  // std::map<std::string nickname, User user> nicknameToUserMap;
-  std::map<std::string, User> nicknameToUserMap;
-  // std::map<std::string channelName, Channel & channel>
-  // channelNameToChannelMap;
-  std::map<std::string, Channel> channelNameToChannelMap;
-  std::set<std::string> _nickHistory;
-
-  Server();
-  Server(const Server &server);
-  Server &operator=(const Server &server);
+public:
 };
 
-#endif
+// User ======================================================
+// status
+#define NOT_REGISTERED 0
+#define PASSWORD_SENT 1
+
+class User {
+private:
+  int _fd;
+  std::string _nick;    // 9文字まで
+  std::string _service; // nick + server_name
+  std::string _buf;
+  std::vector<std::string> _command; //
+  int _status; // statusをdefineしておくといいかも
+  int _level;  // userがどの程度の権限を持っているのかを保持
+  int _mode;
+public:
+  User();
+  ~User();
+  int receive();
+  void validate_input_format();
+  void check_user_status();
+
+  // setter getter
+  void set_fd(const int fd);
+  int get_fd()const;
+  void set_buf(const std::string buf);
+  std::string get_buf()const;
+};
+
+// Channnel ====================================================
+class Channnel {
+private:
+  std::string _name;
+  std::string _topic;
+
+  std::vector<User> _users; // 参加者
+public:
+};
+
+// Server ======================================================
+class Server {
+private:
+  int _port;
+  std::string _password;
+  int _fd;
+  struct sockaddr_in _a_addr;
+  std::vector<struct pollfd> _fds;
+  std::vector<User> _users;
+  std::vector<Channnel> _channnels;
+  const std::string _server_name; // 63文字まで
+
+public:
+  Server();
+  int init();
+  int start();
+  int execute(User &user);
+  int newUser();
+  int get_fd()const;
+  void set_port(const int port);
+  int get_port()const;
+  void set_password(const std::string password);
+  std::string get_password()const;
+  User *find_user_by_fd(const int fd);
+};
+
+#endif // SERVER
